Allow an empty language section in extended header encoding.
Section 3.2.1 of RFC 8187 specifies that the language is optional
in headers using extended encoding. One example given in 3.2.3
looks like:

foo: bar; title*=UTF-8''%c2%a3%20and%20%e2%82%ac%20rates

(the quotes are necessary delimiters, but the language is optional
and in this case missing).

This commit adjusts the relevant regular expression to allow an
empty string in this segment.

https://datatracker.ietf.org/doc/html/rfc8187#section-3.2.1